Fix: min -> milliseconds 로 변경 중

+++ app/src/main/java/com/example/plan_me/ui/timer/TimerSettingListener.kt
@@ -1,7 +1,7 @@
 package com.example.plan_me.ui.timer
 
 interface TimerSettingListener {
-    fun onSettingConfirmed(focusTime: Int)
+    fun onSettingConfirmed(focusTime: Long)
 }
 
 // DialogTimerSettingFragment 내부에서 설정 변경 시 호출 되는 메서드
@@ -12,6 +12,6 @@ fun setOnSettingConfirmedListener(listener: TimerSettingListener) {
 }
 
 // 설정 변경 시 호출
-private fun notifySettingConfirmed(focusTime: Int) {
+private fun notifySettingConfirmed(focusTime: Long) {
     settingConfirmedListener?.onSettingConfirmed(focusTime)
 }
